+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  
 - Issue #144: Add resiliency management
+- Issue #179: Add real measure of reaction time
 
 ### Fixed